test: improve coverage 💯 (#52)
* wip: show coverage report in local

* wip: more test cases

* test: class names

* test: svg image plugin

* test: svg image plugin

* refactor: sessions

* fix: only discover snapshots in supported snapshot files

* fix: remove empty snapshot file

* test: type checking

* test: serializer remove snapshot

* chore: task execute all in lint fix

* test: walk snapshot dirs

* refactor: separate image plugins

* test: plugin raw single image

* refactor: cleaner

* refactor: tasks

* chore: remove unneeded flake8 comment

* chore: task match test pattern

* chore: remove print statement from test

+++ src/syrupy/session.py
@@ -1,4 +1,5 @@
 import os
+from collections import namedtuple
 from gettext import (
     gettext,
     ngettext,
@@ -13,13 +14,27 @@ from .terminal import (
     bold,
     error_style,
     green,
-    yellow,
+    success_style,
+    warning_style,
 )
 
 
 if TYPE_CHECKING:
     from .assertion import SnapshotAssertion
+    from .serializers.base import AbstractSnapshotSerializer  # noqa: F401
     from .types import SnapshotFiles
+
+
+SnapshotGroups = namedtuple(
+    "SnapshotGroups",
+    ["created", "discovered", "failed", "matched", "unused", "updated", "used"],
+)
+
+
+def empty_snapshot_groups() -> "SnapshotGroups":
+    return SnapshotGroups(
+        created={}, discovered={}, failed={}, matched={}, unused={}, updated={}, used={}
+    )
 
 
 class SnapshotSession:
@@ -28,18 +43,22 @@ class SnapshotSession:
         self.base_dir = base_dir
         self.report: List[str] = []
         self._assertions: List["SnapshotAssertion"] = []
+        self._serializers: Dict[str, "AbstractSnapshotSerializer"] = {}
+        self._snapshot_groups = empty_snapshot_groups()
 
     def start(self) -> None:
         self.report = []
         self._assertions = []
+        self._serializers = {}
+        self._snapshot_groups = empty_snapshot_groups()
 
     def finish(self) -> None:
         self._collate_snapshots()
-        n_unused = self._count_snapshots(self._unused_snapshots)
-        n_written = self._count_snapshots(self._created_snapshots)
-        n_updated = self._count_snapshots(self._updated_snapshots)
-        n_failed = self._count_snapshots(self._failed_snapshots)
-        n_passed = self._count_snapshots(self._matched_snapshots)
+        n_unused = self._count_snapshots(self._snapshot_groups.unused)
+        n_written = self._count_snapshots(self._snapshot_groups.created)
+        n_updated = self._count_snapshots(self._snapshot_groups.updated)
+        n_failed = self._count_snapshots(self._snapshot_groups.failed)
+        n_passed = self._count_snapshots(self._snapshot_groups.matched)
 
         self.add_report_line()
 
@@ -47,25 +66,25 @@ class SnapshotSession:
         if n_failed:
             summary_lines += [
                 ngettext(
-                    "{} snapshot failed.", "{} snapshots failed.", n_failed,
+                    "{} snapshot failed.", "{} snapshots failed.", n_failed
                 ).format(error_style(n_failed))
             ]
         if n_passed:
             summary_lines += [
                 ngettext(
-                    "{} snapshot passed.", "{} snapshots passed.", n_passed,
-                ).format(green(bold(n_passed)))
+                    "{} snapshot passed.", "{} snapshots passed.", n_passed
+                ).format(success_style(n_passed))
             ]
         if n_written:
             summary_lines += [
                 ngettext(
-                    "{} snapshot generated.", "{} snapshots generated.", n_written,
+                    "{} snapshot generated.", "{} snapshots generated.", n_written
                 ).format(green(n_written))
             ]
         if n_updated:
             summary_lines += [
                 ngettext(
-                    "{} snapshot updated.", "{} snapshots updated.", n_updated,
+                    "{} snapshot updated.", "{} snapshots updated.", n_updated
                 ).format(green(n_updated))
             ]
         if n_unused:
@@ -75,7 +94,7 @@ class SnapshotSession:
             else:
                 text_singular = "{} snapshot unused."
                 text_plural = "{} snapshots unused."
-            text_count = yellow(bold(n_unused))
+            text_count = warning_style(n_unused)
             summary_lines += [
                 ngettext(text_singular, text_plural, n_unused).format(text_count)
             ]
@@ -85,14 +104,9 @@ class SnapshotSession:
             self.add_report_line()
             if self.update_snapshots:
                 self.remove_unused_snapshots(
-                    self._unused_snapshots,
-                    self._used_snapshots,
-                    self._snapshot_assertions,
+                    self._snapshot_groups.unused, self._snapshot_groups.used
                 )
-                for filepath, snapshots in self._unused_snapshots.items():
-                    count = self._count_snapshots({filepath: snapshots})
-                    if not count:
-                        continue
+                for filepath, snapshots in self._snapshot_groups.unused.items():
                     path_to_file = os.path.relpath(filepath, self.base_dir)
                     deleted_snapshots = ", ".join(map(bold, sorted(snapshots)))
                     self.add_report_line(
@@ -116,56 +130,58 @@ class SnapshotSession:
         self,
         unused_snapshot_files: "SnapshotFiles",
         used_snapshot_files: "SnapshotFiles",
-        snapshot_assertions: Dict[str, "SnapshotAssertion"],
     ) -> None:
         for snapshot_file, unused_snapshots in unused_snapshot_files.items():
-            if snapshot_file not in used_snapshot_files:
-                os.remove(snapshot_file)
-                continue
             for snapshot_name in unused_snapshots:
-                self._snapshot_assertions[snapshot_file].serializer.delete_snapshot(
+                self._serializers[snapshot_file].delete_snapshot(
                     snapshot_file, snapshot_name
                 )
+            if (
+                os.path.exists(snapshot_file)
+                and snapshot_file not in used_snapshot_files
+            ):
+                os.remove(snapshot_file)
 
     def _collate_snapshots(self) -> None:
         """
         Prepare session for snapshot reporting
         """
-        self._used_snapshots: "SnapshotFiles" = {}
-        self._failed_snapshots: "SnapshotFiles" = {}
-        self._created_snapshots: "SnapshotFiles" = {}
-        self._updated_snapshots: "SnapshotFiles" = {}
-        self._matched_snapshots: "SnapshotFiles" = {}
-        self._discovered_snapshots: "SnapshotFiles" = {}
-        self._snapshot_assertions: Dict[str, "SnapshotAssertion"] = {}
         for assertion in self._assertions:
             self._merge_snapshot_files_into(
-                self._discovered_snapshots, assertion.discovered_snapshots
+                self._snapshot_groups.discovered, assertion.discovered_snapshots
             )
+            for filepath in assertion.discovered_snapshots:
+                self._serializers[filepath] = assertion.serializer
             for result in assertion.executions.values():
-                self._snapshot_assertions[result.file] = assertion
                 snapshot_file: "SnapshotFiles" = {result.file: {result.name}}
-                self._merge_snapshot_files_into(self._used_snapshots, snapshot_file)
+                self._merge_snapshot_files_into(
+                    self._snapshot_groups.used, snapshot_file
+                )
                 if result.created:
                     self._merge_snapshot_files_into(
-                        self._created_snapshots, snapshot_file
+                        self._snapshot_groups.created, snapshot_file
                     )
                 elif result.updated:
                     self._merge_snapshot_files_into(
-                        self._updated_snapshots, snapshot_file
+                        self._snapshot_groups.updated, snapshot_file
                     )
                 elif result.success:
                     self._merge_snapshot_files_into(
-                        self._matched_snapshots, snapshot_file
+                        self._snapshot_groups.matched, snapshot_file
                     )
                 else:
                     self._merge_snapshot_files_into(
-                        self._failed_snapshots, snapshot_file
+                        self._snapshot_groups.failed, snapshot_file
                     )
 
-        self._unused_snapshots: "SnapshotFiles" = self._diff_snapshot_files(
-            self._discovered_snapshots, self._used_snapshots
-        )
+        unused_snapshot_files = {
+            snapshot_file: unused_snapshots
+            for snapshot_file, unused_snapshots in self._diff_snapshot_files(
+                self._snapshot_groups.discovered, self._snapshot_groups.used
+            ).items()
+            if unused_snapshots
+        }
+        self._snapshot_groups.unused.update(unused_snapshot_files)
 
     def _merge_snapshot_files_into(
         self, snapshot_files: "SnapshotFiles", *snapshot_files_to_merge: "SnapshotFiles"
